+++ zenml/pipelines/nlp_pipeline.py
@@ -22,15 +22,13 @@ from tfx.components.statistics_gen.component import StatisticsGen
 from tfx.components.trainer.component import Trainer
 from tfx.proto import trainer_pb2
 
-from zenml.core.backends.training import TrainingBaseBackend
-from zenml.core.components.data_gen.component import DataGen
-from zenml.core.components.split_gen.component import SplitGen
-from zenml.core.components.tokenizer.component import Tokenizer
-from zenml.core.pipelines import BasePipeline
-from zenml.core.standards import standard_keys as keys
-from zenml.core.steps.split import BaseSplit
-from zenml.core.steps.tokenizer import BaseTokenizer
-from zenml.core.steps.trainer import BaseTrainerStep
+from zenml.backends.training import TrainingBaseBackend
+from zenml.components import DataGen, SplitGen, Tokenizer
+from zenml.pipelines import BasePipeline
+from zenml.standards import standard_keys as keys
+from zenml.steps.split import BaseSplit
+from zenml.steps.tokenizer import BaseTokenizer
+from zenml.steps.trainer import BaseTrainerStep
 from zenml.utils import constants
 from zenml.utils.enums import GDPComponent, PipelineStatusTypes
